get timesheet

+++ src/database/migrations/1681565472146_timesheet.js
@@ -12,7 +12,7 @@ exports.up = (pgm) => {
       onDelete: 'cascade',
     },
     registred_hour: {
-      type: 'integer',
+      type: 'double',
       notNull: true,
     },
     registred_date: {
